Allow overriding auto_update default value in global plugin config (#2)
It's now possible to set a "default auto_update" in the plugin config UI.
If a material doesn't have an explicit auto_update field in the gocd.yaml config, this
default value is used instead. This allows to override GoCD's default (true).

This change only affects git materials at the moment. It is not relevent for non-scm materials
(e.g. package materials) and we're only concerned with git scm materials at this point.

Besides a global plugin config, I also tried using a config-repo specific
default value (i.e. setting it in cruise.xml in the <config-repo> element).
This doesn't work for (preflight) validation because the server doesn't send this configuration
(and it couldn't anyway if no repo id is given in the preflight request). So validation can fail
because it will always use the default auto_update=true even though the material is already loaded
with auto_update=false.

The change is deliberately made to be minimally intrusive, i.e. without change a large number of files, to
simplify merging upstream changes (if upstream ever becomes active again). Hence the somewhat ugly TransformConfig
singleton.