Make default auto update a global plugin setting
The per-config-repo configuration doesn't work during preflight
validation because the server doesn't send the configuration (and it
couldn't anyway if no repo id is given in the preflight request).
So instead, make the setting a global plugin setting.

Also adjusted the transform behavior so it only affects git materials,
and not for example upstream dependencies.